new: dev: IteratorConcater for joining multiple list of object iterators into one. chg: dev: use dependency on ChunkDao and ResourceMapper instead of double ResourceDa for MultipartUploadDao. fix: dev: conflict on saving new Resource and new Upload in initiateUpload. chg: dev: remove completeUpload dao method. fix: dev: insert SubResource without lacking fields when converting to ChunkedResourceBean in MultipartUploadDaoImpl. new: dev: database query method support related to UploadDao in ResourceMapper.

+++ src/main/java/com/chigix/resserver/mybatis/MultipartUploadDaoImpl.java
@@ -5,17 +5,21 @@ import com.chigix.resserver.entity.Bucket;
 import com.chigix.resserver.entity.ChunkedResource;
 import com.chigix.resserver.entity.MultipartUpload;
 import com.chigix.resserver.entity.dao.MultipartUploadDao;
+import com.chigix.resserver.entity.error.InvalidPart;
 import com.chigix.resserver.entity.error.NoSuchBucket;
-import com.chigix.resserver.entity.error.NoSuchKey;
 import com.chigix.resserver.entity.error.NoSuchUpload;
 import com.chigix.resserver.mybatis.bean.AmassedResourceBean;
 import com.chigix.resserver.mybatis.bean.BucketBean;
 import com.chigix.resserver.mybatis.bean.ChunkedResourceBean;
 import com.chigix.resserver.mybatis.dto.MultipartUploadDto;
+import com.chigix.resserver.mybatis.dto.ResourceBuilder;
+import com.chigix.resserver.mybatis.dto.ResourceDto;
 import java.security.InvalidParameterException;
 import java.util.Iterator;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
 import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
 
 /**
  *
@@ -25,16 +29,15 @@ public class MultipartUploadDaoImpl implements MultipartUploadDao {
 
     private final MultipartUploadMapper uploadMapper;
 
-    private final ResourceDaoImpl uploadedResourceDao;
-
-    private final ResourceDaoImpl uploadingResourceDao;
+    private final ChunkDaoImpl chunkdao;
+    private final ResourceMapper uploadingResourceMapper;
 
     public MultipartUploadDaoImpl(MultipartUploadMapper uploadMapper,
-            ResourceDaoImpl uploadedResourceDao,
-            ResourceDaoImpl uploadingResourceDao) {
+            ChunkDaoImpl chunkdao,
+            ResourceMapper uploadingResourceMapper) {
         this.uploadMapper = uploadMapper;
-        this.uploadingResourceDao = uploadingResourceDao;
-        this.uploadedResourceDao = uploadedResourceDao;
+        this.uploadingResourceMapper = uploadingResourceMapper;
+        this.chunkdao = chunkdao;
     }
 
     @Override
@@ -45,61 +48,77 @@ public class MultipartUploadDaoImpl implements MultipartUploadDao {
         } catch (ClassCastException classCastException) {
             throw new InvalidParameterException("Bucket inside this resource is not a persisted bean object.");
         }
-        uploadingResourceDao.saveResource(resource);
-        MultipartUpload u;
+        final AtomicReference<AmassedResource> resource_ref = new AtomicReference<>();
+        resource_ref.set(resource);
+        MultipartUpload u = new MultipartUpload(resource) {
+            @Override
+            public AmassedResource getResource() {
+                return resource_ref.get();
+            }
+
+        };
+        uploadingResourceMapper.mergeUploadingResource(new ResourceDto(resource, bb), new MultipartUploadDto(u));
+        ResourceBuilder b = uploadingResourceMapper
+                .selectByUploadId(u.getUploadId());
+        AmassedResourceBean resource_bean;
         try {
-            u = new MultipartUpload((AmassedResource) uploadingResourceDao.findResource(bb, resource.getKey(), resource.getVersionId()));
-        } catch (NoSuchKey ex) {
-            throw new RuntimeException("Unexpected Exception that amassedResource should be saved previously.", ex);
+            resource_bean = (AmassedResourceBean) b.build(null, chunkdao, null);
+        } catch (NullPointerException e) {
+            if (b == null) {
+                throw new RuntimeException("Unexpected!!! AmassedResource have "
+                        + "been removed while belonged upload is still initiating. : ["
+                        + resource.getKey() + "]");
+            }
+            throw e;
         }
+        resource_bean.setBucket(bb);
+        resource_ref.set(resource_bean);
         uploadMapper.insert(new MultipartUploadDto(u));
         return u;
     }
 
     @Override
-    public MultipartUpload findUpload(String uploadId) throws NoSuchUpload, NoSuchBucket {
+    public MultipartUpload findUpload(String uploadId) throws NoSuchUpload {
         Map<String, String> row = uploadMapper.selectByUuid(uploadId);
-        MultipartUpload u;
-        try {
-            u = new MultipartUpload((AmassedResource) uploadingResourceDao.findResource(
-                    new BucketBean(row.get("bucket_name"), new DateTime(), row.get("bucket_uuid")),
-                    row.get("resource_key"), row.get("resource_version")),
-                    uploadId, DateTime.parse(row.get("initiated_at")));
-        } catch (NoSuchKey ex) {
+        ResourceBuilder b = uploadingResourceMapper.selectByUploadId(uploadId);
+        if (b == null) {
             throw new NoSuchUpload();
         }
+        AmassedResourceBean resource_bean = (AmassedResourceBean) b.build(null, chunkdao, null);
+        resource_bean.setBucket(new BucketBean(row.get("bucket_name"),
+                new DateTime(DateTimeZone.UTC), row.get("bucket_uuid")));
+        MultipartUpload u = new MultipartUpload(resource_bean, uploadId, DateTime.parse(row.get("initiated_at")));
         return u;
     }
 
     @Override
-    public ChunkedResource findUploadPart(MultipartUpload upload, String partNumber, String etag) {
+    public ChunkedResource findUploadPart(MultipartUpload upload, String partNumber, String etag) throws InvalidPart {
         String key = partNumber;
         if (key.length() > 32) {
             key = key.substring(key.length() - 32);
         }
-        return uploadingResourceDao.findSubResourcePart(key, etag, upload.getResource());
-    }
-
-    @Override
-    public MultipartUpload completeUpload(MultipartUpload upload) throws NoSuchBucket, NoSuchUpload {
-        if (uploadMapper.deleteByUuid(upload.getUploadId()) < 1) {
-            throw new NoSuchUpload();
+        ResourceBuilder b = uploadingResourceMapper.selectSubResourceByKeyEtagParent(
+                key, upload.getResource().getVersionId(), etag);
+        if (b == null) {
+            throw new InvalidPart();
         }
+        ChunkedResourceBean resource_bean = (ChunkedResourceBean) b.build(null, chunkdao, null);
         try {
-            uploadingResourceDao.removeResource(upload.getResource());
-        } catch (NoSuchKey ex) {
-            return null;
+            resource_bean.setBucket((BucketBean) upload.getResource().getBucket());
+        } catch (NoSuchBucket ex) {
+            throw new RuntimeException("Unexpected!!!!AmassedResource's BucketBean object isn't put previously.");
         }
-        uploadedResourceDao.saveResource(upload.getResource());
-        return upload;
+        resource_bean.setParentResource((AmassedResourceBean) upload.getResource());
+        return resource_bean;
     }
 
     @Override
     public void removeUpload(MultipartUpload upload) throws NoSuchUpload {
-        if (uploadMapper.deleteByUuid(upload.getUploadId()) > 0) {
-            return;
+        int upload_count = uploadMapper.deleteByUuid(upload.getUploadId());
+        int resource_count = uploadingResourceMapper.deleteByUploadId(upload.getUploadId());
+        if (upload_count < 1 || resource_count < 1) {
+            throw new NoSuchUpload();
         }
-        throw new NoSuchUpload();
     }
 
     @Override
@@ -120,16 +139,14 @@ public class MultipartUploadDaoImpl implements MultipartUploadDao {
             @Override
             public MultipartUpload next() {
                 Map<String, String> row = rows.next();
-                try {
-                    return new MultipartUpload(
-                            (AmassedResource) uploadingResourceDao.findResource(
-                                    new BucketBean(row.get("bucket_name"), new DateTime(), row.get("bucket_uuid")),
-                                    row.get("resource_key"), row.get("resource_version")),
-                            row.get("uuid"), DateTime.parse(row.get("initiated_at")));
-                } catch (NoSuchKey | NoSuchBucket ex) {
+                ResourceBuilder b = uploadingResourceMapper.selectByUploadId(row.get("uuid"));
+                if (b == null) {
                     uploadMapper.deleteByUuid(row.get("uuid"));
                     return next();
                 }
+                AmassedResourceBean resource_bean = (AmassedResourceBean) b.build(null, chunkdao, null);
+                resource_bean.setBucket(new BucketBean(row.get("bucket_name"), new DateTime(DateTimeZone.UTC), row.get("bucket_uuid")));
+                return new MultipartUpload(resource_bean, row.get("uuid"), DateTime.parse(row.get("initiated_at")));
             }
         };
     }
@@ -144,10 +161,14 @@ public class MultipartUploadDaoImpl implements MultipartUploadDao {
         if (key.length() > 32) {
             key = key.substring(key.length() - 32);
         }
-        ChunkedResourceBean chunk_bean = new ChunkedResourceBean(key, r.getVersionId(), r.getETag());
+        ResourceBuilder chunked_resource_builder = ResourceBuilder.createFrom(
+                new ResourceDto(r, amassed_resource.getBucket()));
+        chunked_resource_builder.setResourceKey(key);
+        ChunkedResourceBean chunk_bean
+                = (ChunkedResourceBean) chunked_resource_builder.build(null, chunkdao, null);
         chunk_bean.setParentResource(amassed_resource);
-        chunk_bean.setBucket(amassed_resource.getBucket());
-        uploadingResourceDao.saveResource(chunk_bean);
+        uploadingResourceMapper.insertSubResource(new ResourceDto(chunk_bean,
+                amassed_resource.getBucket()));
     }
 
 }
